+++ app/src/main/java/ark/cw_dinner/mainpart/orderingfood/OrderingListAdapter.java
@@ -2,9 +2,7 @@ package ark.cw_dinner.mainpart.orderingfood;
 
 import android.content.Context;
 import android.text.Editable;
-import android.text.TextUtils;
 import android.text.TextWatcher;
-import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -150,7 +148,7 @@ public class OrderingListAdapter extends BaseAdapter {
         return listItemView;
     }
 
-    private int getTotalCost(int portionNum, int portionCost){
+    private double getTotalCost(int portionNum, double portionCost){
         return portionCost * portionNum;
     }
